Refactor KML handling and update related imports.
Renamed `_convert_coords_str` to `to_str` for clarity and improved readability. Adjusted references to the renamed method and updated module imports to reflect the reorganization of `LandCandidateService` into the `kml` module.
